more fixes to get the parallel run through - works now but obviously we are doing parallel work N times the number of threads N ...

+++ dmrg/applications/tools/nrdmexpect.cpp
@@ -71,7 +71,7 @@ typedef U1 symm;
 #include "dmrg/utils/DmrgOptions.h"
 #include "dmrg/utils/DmrgParameters.h"
 
-MPO<matrix, symm> rdm1(Model<matrix, symm> const & model, Lattice const & lattice, size_t const & myrank)
+MPO<matrix, symm> rdm1(Model<matrix, symm> const & model, Lattice const & lattice, int const & myrank)
 {
         typedef operator_selector<matrix, symm>::type op_t;
         typedef OPTable<matrix, symm>::tag_type tag_type;
